refactor: move NewPayoutRow component to separate file
